Move `ParallelString` into the string module
It's not really meant to be implemented by anyone else, so it makes
more sense to live near the implementation.
